Experiment: Windowed Output Streams
I implemented a naive windowing function for the linechange ouput, and
it works, but the limitations of line-buffering mean that the user has
to press enter after typing j to move up or down one line.

Also, there isn't a straightforward way to set the correct number of
lines for the window. If the number of visible lines exceeds the
line-height of the terminal, then the output will still overflow the
screen space. I would like to find a way to dynamically set the line
height based on the terminal height, but this as well as
non-line-buffered input may require a library like ncurses, which would
likely be implemented as a NIF.

More research is required.