WIP: Redesign File Collection Logic
The file collection logic is not very well thought out, and it needs a
redesign.

In particular, there needs to be some way to exclude specific files and
directories from the search.

There should also be a way to include entire directories that are not
under the <root_dir>, but the behavior of that inclusion should follow
the collection behavior for the <root_dir>. I.e. the files collected
should have extensions specified in the --extensions switch. This brings
up a design decision regarding the commandline arguments; should the
--include switch handle directories in addition to regular files, or
should the commandline support multiple <root_dir> arguments?

I need to make a decision about this before merging the exclusion logic into
main.